Make frontend code use list of supported extensions
This removes the last bit of duplicated code that had a hard-coded list
of extensions.

Now, if we want to add support for new extensions all we need to do is
create a new class like `mariner.file_formats.CTBFile` and register it
on `mariner.file_formats.utils.EXTENSION_TO_FILE_FORMAT`.