dh-systemd has been removed from Bullseye.
As per https://bugs.debian.org/822670, the dh-systemd package has been merged with the main debhelper package since version 9.20160709

For some time, a transitional dh-systemd package existed which simply depended on debhelper. However, this was removed from bullseye.

The fix should be as simple as depending on debhelper (>= 9.20160709), which available on all supported Debian releases and all supported Ubuntu releases except for xenial.